+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 - Add navigation links to previous and next years
+- Add option to show all months, even if they are empty
 
 ## [0.3.1] - 2025-04-10